LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/conformance-requirements/2d.type.delete-expected.txt

2d.type.delete
window.CanvasRenderingContext2D is Configurable
Defined in "Web IDL" (draft)
Actual output:

PASS window.CanvasRenderingContext2D is Configurable 
